+++ src/tribler-core/tribler_core/logger/logger.py
@@ -11,7 +11,7 @@ LOG_CONFIG_FILENAME = 'logger.yaml'
 logger = logging.getLogger(__name__)
 
 
-# note: this class is used by src/tribler-common/tribler_common/logger.yaml
+# note: this class is used by logger.yaml
 class StdoutFilter(logging.Filter):
     def filter(self, record):
         return record.levelno < logging.ERROR
@@ -30,7 +30,7 @@ def get_logger_config_path():
     if not hasattr(sys, '_MEIPASS'):
         dirname = Path(__file__).absolute().parent
     else:
-        dirname = Path(getattr(sys, '_MEIPASS'), "tribler_source", "tribler_common")
+        dirname = Path(getattr(sys, '_MEIPASS')) / "tribler_source/tribler_core/logger"
     return dirname / LOG_CONFIG_FILENAME
