+++ src/tribler-core/tribler_core/logger/logger_streams.py
