+++ src/tribler-core/tribler_core/logger/tests/test_logger.py
@@ -2,11 +2,11 @@ import logging
 from io import BytesIO, TextIOWrapper
 from unittest.mock import MagicMock, Mock, call, patch
 
-from tribler_common.logger import get_logger_config_path, setup_logging
-from tribler_common.logger_streams import StreamWrapper
+from tribler_core.logger.logger import get_logger_config_path, setup_logging
+from tribler_core.logger.logger_streams import StreamWrapper
 
 
-@patch('tribler_common.logger.__file__', '/a/b/c/logger.py')
+@patch('tribler_core.logger.logger.__file__', '/a/b/c/logger.py')
 def test_get_logger_config_path():
     config_path = get_logger_config_path()
     # take the last part of the path to ignore a drive name on Windows
@@ -14,10 +14,10 @@ def test_get_logger_config_path():
 
     with patch('sys._MEIPASS', '/x/y/z/', create=True):
         config_path = get_logger_config_path()
-        assert config_path.parts[-6:] == ('x', 'y', 'z', 'tribler_source', 'tribler_common', 'logger.yaml')
+        assert config_path.parts[-7:] == ('x', 'y', 'z', 'tribler_source', 'tribler_core', 'logger', 'logger.yaml')
 
 
-@patch('tribler_common.logger.logger')
+@patch('tribler_core.logger.logger.logger')
 @patch('sys.stdout')
 @patch('sys.stderr')
 @patch('builtins.print')
@@ -38,7 +38,7 @@ def test_setup_logging_no_config(basic_config: Mock, print_: Mock, stderr: Mock,
 
 @patch('yaml.safe_load')
 @patch('logging.config.dictConfig')
-@patch('tribler_common.logger.logger')
+@patch('tribler_core.logger.logger.logger')
 def test_setup_logging(logger: Mock, dict_config: Mock, yaml_safe_load: Mock):
     log_dir = MagicMock()
     log_dir.__str__.return_value = '<log-dir>'
@@ -66,7 +66,7 @@ def test_setup_logging(logger: Mock, dict_config: Mock, yaml_safe_load: Mock):
     )
 
 
-@patch('tribler_common.logger.logger')
+@patch('tribler_core.logger.logger.logger')
 @patch('sys.stdout')
 @patch('sys.stderr')
 @patch('builtins.print')
