+++ src/tribler-core/tribler_core/utilities/tests/test_utilities.py
@@ -5,7 +5,7 @@ from aiohttp import ClientSession
 
 import pytest
 
-from tribler_common.logger import load_logger_config
+from tribler_core.logger.logger import load_logger_config
 from tribler_common.patch_import import patch_import
 
 from tribler_core.utilities.tracker_utils import add_url_params
