+++ src/tribler-gui/tribler_gui/start_gui.py
@@ -4,7 +4,7 @@ import sys
 
 from PyQt5.QtCore import QSettings
 
-from tribler_common.logger import load_logger_config
+from tribler_core.logger.logger import load_logger_config
 from tribler_common.rest_utils import path_to_uri
 from tribler_common.sentry_reporter.sentry_reporter import SentryStrategy
